backend/pyproject.toml
[project]
name = "tadabbur-ai"
version = "0.1.0"
description = "RAG-grounded Quranic knowledge platform with story connections"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",

    # Vector Database
    "qdrant-client>=1.7.0",

    # Cache & Queue
    "redis>=5.0.0",
    "rq>=1.16.0",

    # AI/ML
    "anthropic>=0.18.0",
    "sentence-transformers>=2.3.0",
    "torch>=2.1.0",

    # Data Processing
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.1",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
